#add 登录

+++ Neil.Model/ViewModel/UserInfoModel.cs
@@ -12,7 +12,7 @@ namespace Neil.Model.ViewModel
         public string UName { get; set; }
         public string UPwd { get; set; }
         public System.DateTime SubTime { get; set; }
-        public short DelFlag { get; set; }
+        public int DelFlag { get; set; }
         public System.DateTime ModifiedOn { get; set; }
         public string Remark { get; set; }
         public string Sort { get; set; }
